Convert partials to components, and move templates to addon
This converts all partials to components, requiring that some data be passed from the previously shared component to the control view and label cmponents used for each control.  In addition Ember 1.11 attribute syntax is now used, and all templates are in the addon directory and incorporated as layout for each component removing any use of the deprecated layoutName.